Вынес функции из header-view в utils
